+++ src/1-symm/rijndael.h
@@ -21,16 +21,16 @@ void AES256_KeyExpansion(void const *restrict key, void *restrict w);
         q==blockBytes ? 16 :                                    \
         q==keyBytes ? (bits)/8 :                                \
         q==keyschedBytes ? ((bits)/32+6+1)*16 :                 \
-        q==EncFunc ? (uintptr_t)AES##bits##_Cipher :            \
-        q==DecFunc ? (uintptr_t)AES##bits##_InvCipher :         \
-        q==KschdFunc ? (uintptr_t)AES##bits##_KeyExpansion :    \
+        q==EncFunc ? (uintmax_t)AES##bits##_Cipher :            \
+        q==DecFunc ? (uintmax_t)AES##bits##_InvCipher :         \
+        q==KschdFunc ? (uintmax_t)AES##bits##_KeyExpansion :    \
         0)
 #define cAES128(q) cAES(128,q)
 #define cAES192(q) cAES(192,q)
 #define cAES256(q) cAES(256,q)
 
-uintptr_t iAES128(int q);
-uintptr_t iAES192(int q);
-uintptr_t iAES256(int q);
+uintmax_t iAES128(int q);
+uintmax_t iAES192(int q);
+uintmax_t iAES256(int q);
 
 #endif /* MySuiteA_rijndael_h */
